+++ tomcat/src/main/java/com/techcourse/controller/RegisterController.java
@@ -6,9 +6,8 @@ import com.techcourse.util.StaticResourceManager;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Map;
-import org.apache.catalina.Manager;
 import org.apache.catalina.controller.AbstractController;
-import org.apache.coyote.http11.HttpStatusCode;
+import org.apache.coyote.http11.common.HttpStatusCode;
 import org.apache.coyote.http11.response.HttpResponse;
 import org.apache.coyote.MediaType;
 import org.apache.coyote.http11.request.HttpRequest;
