+++ tomcat/src/main/java/org/apache/coyote/http11/common/HttpMethod.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.common;
 
 import java.util.HashMap;
 import java.util.Map;
